requirements.txt
azure-ai-ml>=1.12.0
azure-identity>=1.15.0
mlflow>=2.9.0,<3.0.0
azureml-mlflow>=1.57.0
pandas>=2.1.0
scikit-learn>=1.3.0
matplotlib>=3.8.0
seaborn>=0.13.0
numpy>=1.26.0
faker>=22.0.0
